GG-23666 Packages structure fixed for OpenCensus SPIs
