+++ modules/opencensus/src/main/java/org/apache/ignite/spi/metric/opencensus/package-info.java
@@ -14,20 +14,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.opencensus.spi;
-
-import org.apache.ignite.opencensus.spi.metric.OpenCensusMetricExporterSpiTest;
-import org.apache.ignite.opencensus.spi.tracing.OpenCensusTracingSpiTest;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-    OpenCensusMetricExporterSpiTest.class,
-    OpenCensusTracingSpiTest.class
-})
 /**
- * Suite to test OpenCensus integration.
+ *
  */
-public class IgniteOpenCensusSuite {
-}
+package org.apache.ignite.spi.metric.opencensus;
+
+/**
+ * <!-- Package description. -->
+ * Contains Opencensus based implementation of MetricExporterSpi.
+ */
