+++ modules/opencensus/src/test/java/org/apache/ignite/internal/processors/monitoring/opencensus/OpenCensusMetricExporterSpiTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.opencensus.spi.metric;
+package org.apache.ignite.internal.processors.monitoring.opencensus;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -32,11 +32,12 @@ import org.apache.ignite.configuration.IgniteConfiguration;
 import org.apache.ignite.internal.IgniteEx;
 import org.apache.ignite.internal.metric.AbstractExporterSpiTest;
 import org.apache.ignite.internal.util.GridConcurrentHashSet;
+import org.apache.ignite.spi.metric.opencensus.OpenCensusMetricExporterSpi;
 import org.junit.Test;
 
-import static org.apache.ignite.opencensus.spi.metric.OpenCensusMetricExporterSpi.CONSISTENT_ID_TAG;
-import static org.apache.ignite.opencensus.spi.metric.OpenCensusMetricExporterSpi.INSTANCE_NAME_TAG;
-import static org.apache.ignite.opencensus.spi.metric.OpenCensusMetricExporterSpi.NODE_ID_TAG;
+import static org.apache.ignite.spi.metric.opencensus.OpenCensusMetricExporterSpi.CONSISTENT_ID_TAG;
+import static org.apache.ignite.spi.metric.opencensus.OpenCensusMetricExporterSpi.INSTANCE_NAME_TAG;
+import static org.apache.ignite.spi.metric.opencensus.OpenCensusMetricExporterSpi.NODE_ID_TAG;
 import static org.apache.ignite.testframework.GridTestUtils.waitForCondition;
 
 /** */
